Move Finder & Locator under the FileSystem namespace
