+++ src/Command/InfectionCommand.php
@@ -45,8 +45,8 @@ use Infection\Console\LogVerbosity;
 use Infection\Container;
 use Infection\Engine;
 use Infection\Event\ApplicationExecutionStarted;
-use Infection\Locator\FileOrDirectoryNotFound;
-use Infection\Locator\Locator;
+use Infection\FileSystem\Locator\FileOrDirectoryNotFound;
+use Infection\FileSystem\Locator\Locator;
 use Infection\TestFramework\Coverage\CoverageDoesNotExistException;
 use Infection\TestFramework\TestFrameworkTypes;
 use function is_numeric;
