+++ src/Config/ValueProvider/ExcludeDirsProvider.php
@@ -38,8 +38,8 @@ namespace Infection\Config\ValueProvider;
 use function count;
 use function in_array;
 use Infection\Config\ConsoleHelper;
-use Infection\Locator\Locator;
-use Infection\Locator\RootsFileOrDirectoryLocator;
+use Infection\FileSystem\Locator\Locator;
+use Infection\FileSystem\Locator\RootsFileOrDirectoryLocator;
 use function Safe\glob;
 use Symfony\Component\Console\Helper\QuestionHelper;
 use Symfony\Component\Console\Input\InputInterface;
