+++ src/Config/ValueProvider/PhpUnitCustomExecutablePathProvider.php
@@ -38,8 +38,8 @@ namespace Infection\Config\ValueProvider;
 use Closure;
 use const DIRECTORY_SEPARATOR;
 use Infection\Config\ConsoleHelper;
-use Infection\Finder\Exception\FinderException;
-use Infection\Finder\TestFrameworkFinder;
+use Infection\FileSystem\Finder\Exception\FinderException;
+use Infection\FileSystem\Finder\TestFrameworkFinder;
 use RuntimeException;
 use Symfony\Component\Console\Helper\QuestionHelper;
 use Symfony\Component\Console\Input\InputInterface;
