+++ tests/phpunit/AutoReview/ProjectCode/ProjectCodeProvider.php
@@ -52,9 +52,9 @@ use Infection\Console\Util\PhpProcess;
 use Infection\Engine;
 use Infection\Event\Subscriber\MutantCreatingConsoleLoggerSubscriber;
 use Infection\Event\Subscriber\MutationGeneratingConsoleLoggerSubscriber;
-use Infection\Finder\ComposerExecutableFinder;
-use Infection\Finder\FilterableFinder;
-use Infection\Finder\TestFrameworkFinder;
+use Infection\FileSystem\Finder\ComposerExecutableFinder;
+use Infection\FileSystem\Finder\FilterableFinder;
+use Infection\FileSystem\Finder\TestFrameworkFinder;
 use Infection\Http\BadgeApiClient;
 use Infection\Logger\ResultsLoggerTypes;
 use Infection\Mutant\MetricsCalculator;
