+++ tests/phpunit/Config/ValueProvider/PhpUnitCustomExecutablePathProviderTest.php
@@ -37,8 +37,8 @@ namespace Infection\Tests\Config\ValueProvider;
 
 use Infection\Config\ConsoleHelper;
 use Infection\Config\ValueProvider\PhpUnitCustomExecutablePathProvider;
-use Infection\Finder\Exception\FinderException;
-use Infection\Finder\TestFrameworkFinder;
+use Infection\FileSystem\Finder\Exception\FinderException;
+use Infection\FileSystem\Finder\TestFrameworkFinder;
 use function Infection\Tests\normalizePath as p;
 use PHPUnit\Framework\MockObject\MockObject;
 use Symfony\Component\Console\Exception\RuntimeException as SymfonyRuntimeException;
