+++ tests/phpunit/Configuration/Schema/SchemaConfigurationLoaderTest.php
@@ -39,7 +39,7 @@ use Generator;
 use Infection\Configuration\Schema\SchemaConfiguration;
 use Infection\Configuration\Schema\SchemaConfigurationFileLoader;
 use Infection\Configuration\Schema\SchemaConfigurationLoader;
-use Infection\Locator\Locator;
+use Infection\FileSystem\Locator\Locator;
 use PHPUnit\Framework\MockObject\MockObject;
 use PHPUnit\Framework\TestCase;
 use ReflectionClass;
