+++ tests/phpunit/Console/E2ETest.php
@@ -43,8 +43,8 @@ use Generator;
 use Infection\Command\ConfigureCommand;
 use Infection\Console\Application;
 use Infection\Container;
-use Infection\Finder\ComposerExecutableFinder;
-use Infection\Finder\Exception\FinderException;
+use Infection\FileSystem\Finder\ComposerExecutableFinder;
+use Infection\FileSystem\Finder\Exception\FinderException;
 use const PHP_SAPI;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Console\Input\ArgvInput;
