+++ tests/phpunit/TestFramework/Config/TestFrameworkConfigLocatorTest.php
@@ -35,7 +35,7 @@ declare(strict_types=1);
 
 namespace Infection\Tests\TestFramework\Config;
 
-use Infection\Locator\FileOrDirectoryNotFound;
+use Infection\FileSystem\Locator\FileOrDirectoryNotFound;
 use Infection\TestFramework\Config\TestFrameworkConfigLocator;
 use function Infection\Tests\normalizePath as p;
 use PHPUnit\Framework\TestCase;
